function that returns length of list

+++ 0x13-more_singly_linked_lists/1-listint_len.c
@@ -6,7 +6,7 @@
  */
 size_t listint_len(const listint_t *h)
 {
-	unsigned int count = 0;
+	size_t int count = 0;
 	const listint_t *current;
 
 	current = h;
